requirements.txt
praw
google-generativeai
streamlit